app/requirements.txt
asgiref==3.7.2
Django==4.2.6
djangorestframework==3.14.0
dnspython==2.4.2
email-validator==2.0.0.post2
idna==3.4
phonenumbers==8.13.22
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
validate-docbr==1.10.0
